bundler: Rename keywords
